painfully upgraded to enw VisUI... for the reason of upgrading to libGDX 1.7.1

+++ overlap2d/src/com/uwsoft/editor/view/ui/properties/panels/UISpriteAnimationItemProperties.java
@@ -79,7 +79,7 @@ public class UISpriteAnimationItemProperties extends UIItemCollapsibleProperties
     }
 
     public int getFPS() {
-        return fpsSelector.getValue();
+        return (int) fpsSelector.getValue();
     }
 
     public void setPlayModes() {
